backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Database
motor>=3.4.0
pymongo>=4.7.0

# Auth
PyJWT>=2.8.0
bcrypt>=4.1.3

# HTTP
httpx>=0.27.0

# ML — core (NO external API needed)
spacy>=3.8.0
sentence-transformers>=3.0.1
scikit-learn>=1.5.0
numpy>=1.26.0

# Optional: PDF export (comment out if GTK not available on Windows)
# weasyprint>=62.3

# Dev
python-dotenv>=1.0.1
